Added support for <passfocus> message, to launch/switch to another app.
For example:

<passfocus 18>
application: clock

will launch the "clock" application (or switch to it if it is already
running). It will have focus.

<closing 0>

tells Honcho that the application is closing. The application does not
have focus and is removed from the ApplicationStack. The previous
application that had focus now has focus.

The apps/ directory now contains a "clock" application that is, again,
there to help test all parts of Honcho. Selecting "clock" from the
main_menu program will launch the app. It will render the time every
second. Selecting "back" from within the clock app closes it.

Also in this commit is a bunch of refactoring of the Message class.